test transfer functions gradient
